Unsupervised Clustering of Texts
In this projects, I intend to explore different unsupervised clustering
techniques to explore topics metioned in a series of documents. The goal
of the projects is to tease out topics and visulize them in a
human-friendly way. I also intend to construct a dashboard (using
Rshiny) for data exploration and presentation.